refactor: split admin_api and superuser_api
adds dependencies `check_admin` and `check_super_user` to APIRouter.
goal of this pr is to make it clearer which api endpoint has which permissions.
