+++ lnbits/core/templates/admin/index.html
@@ -537,7 +537,7 @@
       },
       restartServer() {
         LNbits.api
-          .request('GET', '/admin/api/v1/restart/?usr=' + this.g.user.id)
+          .request('GET', '/superuser/api/v1/restart/?usr=' + this.g.user.id)
           .then(response => {
             this.$q.notify({
               type: 'positive',
@@ -554,7 +554,7 @@
         LNbits.api
           .request(
             'PUT',
-            '/admin/api/v1/topup/?usr=' + this.g.user.id,
+            '/superuser/api/v1/topup/?usr=' + this.g.user.id,
             this.g.user.wallets[0].adminkey,
             this.wallet
           )
@@ -675,7 +675,7 @@
             LNbits.api
               .request(
                 'DELETE',
-                '/admin/api/v1/settings/?usr=' + this.g.user.id
+                '/superuser/api/v1/settings/?usr=' + this.g.user.id
               )
               .then(response => {
                 this.$q.notify({
@@ -692,7 +692,7 @@
           })
       },
       downloadBackup() {
-        window.open('/admin/api/v1/backup/?usr=' + this.g.user.id, '_blank')
+        window.open('/superuser/api/v1/backup/?usr=' + this.g.user.id, '_blank')
       }
     }
   })
